# doc/api.yml
openapi: 3.0.3
info:
  version: "latest"
  title: Prehakaton - API
  description: |
    XYXYXY
security:
  - BearerAuth: []
paths:
...
  "/users":
    get:
      tags:
        - User
      summary: Get all users.
      operationId: getUsers
      parameters:
        - name: page
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Page"
        - name: page_size
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/PageSize"
        - name: name
          in: query
          required: false
          description: Filter user by first name or lastname or username.
          schema:
            type: string
      responses:
        "200":
          description: List of users that meet the criterias.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  "/users/{id}":
    get:
      tags:
        - User
      summary: Get user by identifier.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      operationId: getUserById
      responses:
        "200":
          description: The identified user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
    put:
      tags:
        - User
      summary: Crupdate user by identifier.
      operationId: crupdateUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: User to crupdate.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: The crupdated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /users/{id}/interested/courses:
    get:
      summary: "Récupérer les cours marqués comme intéressants par l'utilisateur"
      operationId: "getUserInterestedCourses"
      tags:
        - User
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: "ID de l'utilisateur"
      responses:
        '200':
          description: "Liste des cours marqués comme intéressants"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
        '400':
          $ref: "#/components/responses/400"
        '403':
          $ref: "#/components/responses/403"
        '404':
          $ref: "#/components/responses/404"
        '429':
          $ref: "#/components/responses/429"
        '500':
          $ref: "#/components/responses/500"
  /users/{id}/subscribe/courses:
    get:
      summary: "Récupérer les cours suivis par l'utilisateur"
      operationId: "getUserSubscribedCourses"
      tags:
        - User
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: "ID de l'utilisateur"
      responses:
        '200':
          description: "Liste des cours suivis"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
        '400':
          $ref: "#/components/responses/400"
        '403':
          $ref: "#/components/responses/403"
        '404':
          $ref: "#/components/responses/404"
        '429':
          $ref: "#/components/responses/429"
        '500':
          $ref: "#/components/responses/500"
...
  /users/{id}/courses/{cid}/follow/{follow_action}:
    put:
      summary: "Mettre à jour l'état de suivi ou d'intérêt d'un cours pour l'utilisateur"
      operationId: "putUserCourseFollow"
      tags:
        - Courseing
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: "ID de l'utilisateur"
        - in: path
          name: cid
          required: true
          schema:
            type: string
          description: "ID du cours"
        - in: path
          name: follow_action
          required: true
          schema:
            $ref: "#/components/schemas/ActionEnum"
          description: "Action à effectuer sur le cours"
      responses:
        '200':
          description: "État du cours mis à jour"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          $ref: "#/components/responses/400"
        '403':
          $ref: "#/components/responses/403"
        '404':
          $ref: "#/components/responses/404"
        '429':
          $ref: "#/components/responses/429"
        '500':
          $ref: "#/components/responses/500"


















  ## CourseSessioning resources
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  responses:
...
    "400":
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BadRequestException"
    "403":
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NotAuthorizedException"
    "404":
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ResourceNotFoundException"
    "429":
      description: Too many requests to the API
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TooManyRequestsException"
    "500":
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/InternalServerException"
  schemas:
...
    Page:
      type: integer
      description: Start at `1`
    PageSize:
      type: integer
      description: Cannot exceed `500`
    Role:
      type: string
      enum:
        - CLIENT
        - MANAGER
    Sex:
      type: string
      enum:
        - M
        - F
        - OTHER
...
    UserStatus:
      type: string
      enum:
        - ENABLED
        - BANISHED
    ActionEnum:
      type: string
      enum:
        - FOLLOW
        - INTERESTED
        - UNFOLLOW
        - UNINTERESTED
    StatusEnum:
      type: string
      enum:
        - planning
        - confirmed
        - in_progress
        - completed

        ## resource model
    ReactionsProperties:
      properties:
        like_number:
          type: integer
        view_number:
          type: integer
        star_medium:
          type: number
          format: float
    HaveImageProperties:
      properties:
        image_url:
          type: string
        picture_is_implemented:
          type: boolean
    Course:
      allOf:
        - $ref: "#/components/schemas/ReactionsProperties"
        - $ref: "#/components/schemas/HaveImageProperties"
      properties:
        id:
          type: string
        creation_datetime:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
        title:
          type: string
        description:
          type: string
        categories:
          type: array
          items:
            $ref: "#/components/schemas/CourseCategory"
        followers:
          type: array
          items:
            $ref: "#/components/schemas/User"
        interested_users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/CourseSession"
        status:
          $ref: "#/components/schemas/StatusEnum"
        price:
          type: number
          format: float
...
    CourseSession:
      allOf:
        - $ref: "#/components/schemas/ReactionsProperties"
      properties:
        id:
          type: string
        creation_datetime:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        course:
          $ref: "#/components/schemas/Course"
        location:
          $ref: "#/components/schemas/Location"
        professor:
          $ref: "#/components/schemas/User"
        title:
          type: string
        description:
          type: string
        start_datetime:
          type: string
          format: date-time
        end_datetime:
          type: string
          format: date-time
    Location:
      properties:
        id:
          type: string
        creation_datetime:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        name:
          type: string
        description:
          type: string
        latitude:
          type: string
        longitude:
          type: string
...
    CreateUser:
      properties:
        id:
          type: string
        last_name:
          type: string
        first_name:
          type: string
        birth_date:
          type: string
          format: date
        email:
          type: string
        username:
          type: string
        status:
          $ref: "#/components/schemas/UserStatus"
        sex:
          $ref: "#/components/schemas/Sex"
        role:
          $ref: "#/components/schemas/Role"
    User:
      allOf:
        - $ref: "#/components/schemas/CreateUser"
      properties:
        photo_url:
          type: string
        profile_banner_url:
          type: string
        entrance_datetime:
          type: string
          format: date-time
...
    CourseCategory:
      type: object
      properties:
        name:
          type: string
          description: "Nom de la categorie"
        description:
          type: string
          description: "Détails supplémentaires concernant la categorie"

    ## exception model
    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    BadRequestException:
      allOf:
        - $ref: "#/components/schemas/Exception"
      example:
        type: BadRequestException
        message: Bad request
    NotAuthorizedException:
      allOf:
        - $ref: "#/components/schemas/Exception"
      example:
        type: NotAuthorizedException
        message: Not authorized
    ResourceNotFoundException:
      allOf:
        - $ref: "#/components/schemas/Exception"
      example:
        type: ResourceNotFoundException
        message: Resource of type <T> identified by <I> not found
    TooManyRequestsException:
      allOf:
        - $ref: "#/components/schemas/Exception"
      example:
        type: TooManyRequestsException
        message: Too many requests
    InternalServerException:
      allOf:
        - $ref: "#/components/schemas/Exception"
      example:
        type: InternalServerException
        message: Unexpected error